fix(indicator): demo拆解与规范
