+++ src/packages/indicator/demos/h5/demo1.tsx
@@ -1,0 +1,19 @@
+import React from 'react'
+import { Indicator, Cell } from '@nutui/nutui-react'
+
+const Demo1 = () => {
+  return (
+    <>
+      <Cell>
+        <Indicator total={3} current={0} />
+      </Cell>
+      <Cell>
+        <Indicator total={3} current={1} />
+      </Cell>
+      <Cell>
+        <Indicator total={3} current={2} />
+      </Cell>
+    </>
+  )
+}
+export default Demo1
