+++ src/packages/indicator/demos/h5/demo3.tsx
@@ -1,0 +1,20 @@
+import React from 'react'
+import { Indicator, Cell, ConfigProvider } from '@nutui/nutui-react'
+
+const customTheme = {
+  nutuiIndicatorColor: '#3768fa',
+  nutuiIndicatorDotColor: '#ddd',
+  nutuiIndicatorDotSize: '8px',
+  nutuiIndicatorDotActiveSize: '24px',
+}
+
+const Demo3 = () => {
+  return (
+    <Cell>
+      <ConfigProvider theme={customTheme}>
+        <Indicator total={6} current={5} />
+      </ConfigProvider>
+    </Cell>
+  )
+}
+export default Demo3
